move types and fix errors
